User Cart

+++ backend/src/main/java/com/example/backend/dto/DishDTO.java
@@ -11,7 +11,7 @@ public class DishDTO {
     private DishCategoryEnum category;
     private String name;
     private String description;
-    private float price;
+    private BigDecimal price;
     private boolean vegetarian;
     private boolean availability;
 }
